refactor: domain 패키지 내에서 model과 port로 분리
