--- src/main/kotlin/com/server/dpmcore/authority/domain/AuthorityRepository.kt
@@ -1,3 +1,0 @@
-package com.server.dpmcore.authority.domain
-
-interface AuthorityRepository
