+++ src/main/kotlin/com/server/dpmcore/authority/domain/port/AuthorityRepository.kt
@@ -1,0 +1,3 @@
+package com.server.dpmcore.authority.domain.port
+
+interface AuthorityRepository
