--- src/main/kotlin/com/server/dpmcore/cohort/domain/CohortRepository.kt
@@ -1,3 +1,0 @@
-package com.server.dpmcore.cohort.domain
-
-interface CohortRepository
