+++ src/main/kotlin/com/server/dpmcore/cohort/domain/port/CohortRepository.kt
@@ -1,0 +1,3 @@
+package com.server.dpmcore.cohort.domain.port
+
+interface CohortRepository
