--- src/main/kotlin/com/server/dpmcore/member/member/domain/MemberRepository.kt
@@ -1,3 +1,0 @@
-package com.server.dpmcore.member.member.domain
-
-interface MemberRepository
