+++ src/main/kotlin/com/server/dpmcore/member/member/domain/model/MemberId.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.member.member.domain
+package com.server.dpmcore.member.member.domain.model
 
 import java.util.UUID
