+++ src/main/kotlin/com/server/dpmcore/member/member/domain/port/MemberRepository.kt
@@ -1,0 +1,3 @@
+package com.server.dpmcore.member.member.domain.port
+
+interface MemberRepository
