+++ src/main/kotlin/com/server/dpmcore/member/memberCohort/domain/MemberCohort.kt
@@ -1,7 +1,7 @@
 package com.server.dpmcore.member.memberCohort.domain
 
-import com.server.dpmcore.cohort.domain.CohortId
-import com.server.dpmcore.member.member.domain.MemberId
+import com.server.dpmcore.cohort.domain.model.CohortId
+import com.server.dpmcore.member.member.domain.model.MemberId
 
 data class MemberCohort(
     val id: MemberCohortId = MemberCohortId.generate(),
