--- src/main/kotlin/com/server/dpmcore/team/domain/TeamRepository.kt
@@ -1,3 +1,0 @@
-package com.server.dpmcore.team.domain
-
-interface TeamRepository
