+++ src/main/kotlin/com/server/dpmcore/team/domain/model/TeamId.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.team.domain
+package com.server.dpmcore.team.domain.model
 
 import java.util.UUID
