+++ src/main/kotlin/com/server/dpmcore/team/domain/port/TeamRepository.kt
@@ -1,0 +1,3 @@
+package com.server.dpmcore.team.domain.port
+
+interface TeamRepository
